requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
asyncpg>=0.29.0
databricks-sdk>=0.49.0
